LayoutTests/media/video-playsinline.html

<!DOCTYPE html>
<html>
<head>
<script src="media-file.js"></script>
<script src="video-test.js"></script>
<script>
function go()
{
    runNextTest();
}

function runNextTest()
{
    if (tests.length)
        tests.shift()();
    else
        endTest();
}

function testPlaysInline(requiresPlaysInline, requiresWebKitPlaysInline, hasPlaysInline, hasWebKitPlaysInline, expectedDisplayingFullscreen)
{
    run('internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(' + requiresPlaysInline + ')');
    run('internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(' + requiresWebKitPlaysInline + ')');
    video = media = document.createElement('video');
    document.body.appendChild(video);
    video.controls = true;
    video.src = findMediaFile('video', 'content/test');
    run('video.playsInline = ' + hasPlaysInline);
    run('video.' + (hasWebKitPlaysInline ? 'set' : 'remove') + 'Attribute("webkit-playsinline", "")');
    waitForEventOnce('canplaythrough', () => {
        runWithKeyDown(() => {
            run('video.play()');
        });
    })
    waitForEventOnce('playing', () => {
        testExpected('video.webkitDisplayingFullscreen', expectedDisplayingFullscreen);
        document.body.removeChild(video);
        runNextTest();
    });
}

function testRequiresPlaysInlineAndDoesntHaveIt()
{
    testPlaysInline(true, true, false, false, true);
}

function testRequiresPlaysInlineAndDoesntHaveIt2()
{
    testPlaysInline(true, true, false, true, false);
}

function testRequiresPlaysInlineAndDoesntHaveIt3()
{
    testPlaysInline(true, false, false, true, true);
}

function testRequiresPlaysInlineAndDoesntHaveIt4()
{
    testPlaysInline(true, false, false, false, true);
}

function testRequiresPlaysInlineAndDoesHaveIt()
{
    testPlaysInline(true, true, true, true, false);
}

function testRequiresPlaysInlineAndDoesHaveIt2()
{
    testPlaysInline(true, true, true, false, false);
}

function testRequiresPlaysInlineAndDoesHaveIt3()
{
    testPlaysInline(true, false, true, true, false);
}

function testRequiresPlaysInlineAndDoesHaveIt4()
{
    testPlaysInline(true, false, true, false, false);
}

function testDoesNotRequirePlaysInlineAndDoesntHaveIt()
{
    testPlaysInline(false, true, false, true, false);
}

function testDoesNotRequirePlaysInlineAndDoesntHaveIt2()
{
    testPlaysInline(false, true, false, false, true);
}

function testDoesNotRequirePlaysInlineAndDoesntHaveIt3()
{
    testPlaysInline(false, false, false, true, false);
}

function testDoesNotRequirePlaysInlineAndDoesntHaveIt4()
{
    testPlaysInline(false, false, false, false, false);
}

function testDoesNotRequirePlaysInlineAndDoesHaveIt()
{
    testPlaysInline(false, true, true, true, false);
}

function testDoesNotRequirePlaysInlineAndDoesHaveIt2()
{
    testPlaysInline(false, true, true, false, true);
}

function testDoesNotRequirePlaysInlineAndDoesHaveIt3()
{
    testPlaysInline(false, false, true, true, false);
}

function testDoesNotRequirePlaysInlineAndDoesHaveIt4()
{
    testPlaysInline(false, false, true, false, false);
}

var tests = [
    testRequiresPlaysInlineAndDoesntHaveIt,
    testRequiresPlaysInlineAndDoesntHaveIt2,
    testRequiresPlaysInlineAndDoesntHaveIt3,
    testRequiresPlaysInlineAndDoesntHaveIt4,
    testRequiresPlaysInlineAndDoesHaveIt,
    testRequiresPlaysInlineAndDoesHaveIt2,
    testRequiresPlaysInlineAndDoesHaveIt3,
    testRequiresPlaysInlineAndDoesHaveIt4,
    testDoesNotRequirePlaysInlineAndDoesntHaveIt,
    testDoesNotRequirePlaysInlineAndDoesntHaveIt2,
    testDoesNotRequirePlaysInlineAndDoesntHaveIt3,
    testDoesNotRequirePlaysInlineAndDoesntHaveIt4,
    testDoesNotRequirePlaysInlineAndDoesHaveIt,
    testDoesNotRequirePlaysInlineAndDoesHaveIt2,
    testDoesNotRequirePlaysInlineAndDoesHaveIt3,
    testDoesNotRequirePlaysInlineAndDoesHaveIt4
];

</script>
</head>
<body onload="go()">
</body>
</html>
